Implement: combine login into navbar
