+++ src/components/auth/login/GoogleAndFacebookLogIn.jsx
@@ -2,6 +2,9 @@ import React from "react";
 import db from "../../config/firebaseConfig";
 import * as firebase from "firebase";
 import "firebase/auth";
+import Button from "react-bootstrap/Button";
+import Col from "react-bootstrap/Col";
+import Row from "react-bootstrap/Row";
 
 export default function GoogleAndFacebookLogIn() {
   const googleProvider = new firebase.auth.GoogleAuthProvider();
@@ -38,11 +41,27 @@ export default function GoogleAndFacebookLogIn() {
   };
 
   return (
-    <div>
-      <button onClick={() => logIn(googleProvider)}>Log in with Google</button>
-      <button onClick={() => logIn(facebookProvider)}>
-        Log in with Facebook
-      </button>
-    </div>
+    <Row className="socialLoginButtons">
+      <Col>
+        <Button
+          variant="info"
+          size="md"
+          className="fbButton"
+          onClick={() => logIn(facebookProvider)}
+        >
+          Facebook
+        </Button>
+      </Col>
+      <Col>
+        <Button
+          variant="info"
+          size="md"
+          className="googleButton"
+          onClick={() => logIn(googleProvider)}
+        >
+          Google
+        </Button>
+      </Col>
+    </Row>
   );
 }
