+++ src/components/home/Home.jsx
@@ -1,13 +1,11 @@
 import React from "react";
 import Products from "../products/Products";
-import LoginForm from "../auth/login/Login";
 import LanguageDropdown from "./LanguageDropdown";
 import Categories from "./Categories";
 
 export default function Home() {
   return (
     <>
-      <LoginForm />
       <LanguageDropdown />
       <Categories />
       <Products />
